Move asshape() to shapetools.py
