+++ ndindex/array.py
@@ -1,6 +1,7 @@
 import warnings
 
-from .ndindex import NDIndex, asshape
+from .ndindex import NDIndex
+from .shapetools import asshape
 
 class ArrayIndex(NDIndex):
     """
