+++ ndindex/slice.py
@@ -1,5 +1,6 @@
-from .ndindex import NDIndex, asshape, operator_index
+from .ndindex import NDIndex, operator_index
 from .subindex_helpers import subindex_slice
+from .shapetools import asshape
 
 class default:
     """
